# specs/1-obe-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: OBE System API
  description: |
    Outcome-Based Education Management System API
    
    This API provides endpoints for:
    - Authentication (JWT, Google OAuth, Email OTP)
    - User management with RBAC
    - Course and outcome management
    - Assessment and result tracking
    - Analytics and reporting
  version: 1.0.0
  contact:
    name: OBE System API Support
    email: support@obesystem.com
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://api.obesystem.com/api
    description: Production server
...
paths:
...
  /courses:
    get:
      tags: [Courses]
      summary: List courses (role-based filtering)
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResponse'

    post:
      tags: [Courses]
      summary: Create course (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseRequest'
      responses:
        '201':
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /courses/{id}:
    get:
      tags: [Courses]
      summary: Get course details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Course details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Courses]
      summary: Update course (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCourseRequest'
      responses:
        '200':
          description: Course updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Courses]
      summary: Delete course (Admin only, soft delete)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Course deleted successfully
        '403':
          $ref: '#/components/responses/Forbidden'
  /courses/{id}/teachers:
    post:
      tags: [Courses]
      summary: Assign teacher to course (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teacher_id]
              properties:
                teacher_id:
                  type: integer
      responses:
        '201':
          description: Teacher assigned successfully
        '403':
          $ref: '#/components/responses/Forbidden'
  /courses/{id}/students:
    get:
      tags: [Courses]
      summary: List enrolled students (Teacher/Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of enrolled students
          content:
            application/json:
              schema:
                type: object
                properties:
                  students:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

  # Outcome Endpoints
  /plos:
    get:
      tags: [Outcomes]
      summary: List Program Learning Outcomes
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of PLOs
          content:
            application/json:
              schema:
                type: object
                properties:
                  plos:
                    type: array
                    items:
                      $ref: '#/components/schemas/PLOResponse'

    post:
      tags: [Outcomes]
      summary: Create PLO (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePLORequest'
      responses:
        '201':
          description: PLO created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PLOResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /plos/{id}:
    put:
      tags: [Outcomes]
      summary: Update PLO (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePLORequest'
      responses:
        '200':
          description: PLO updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PLOResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Outcomes]
      summary: Delete PLO (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: PLO deleted successfully
        '403':
          $ref: '#/components/responses/Forbidden'
  /courses/{id}/clos:
    get:
      tags: [Outcomes]
      summary: Get CLOs for course
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of CLOs
          content:
            application/json:
              schema:
                type: object
                properties:
                  clos:
                    type: array
                    items:
                      $ref: '#/components/schemas/CLOResponse'

    post:
      tags: [Outcomes]
      summary: Create CLO (Admin/Teacher)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCLORequest'
      responses:
        '201':
          description: CLO created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CLOResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /clos/{id}:
    put:
      tags: [Outcomes]
      summary: Update CLO (Admin/Teacher)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCLORequest'
      responses:
        '200':
          description: CLO updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CLOResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Outcomes]
      summary: Delete CLO (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: CLO deleted successfully
        '403':
          $ref: '#/components/responses/Forbidden'

  # Mapping Endpoints
...
  /analytics/course/{id}:
    get:
      tags: [Analytics]
      summary: Course analytics (Teacher/Admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Course analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseAnalyticsResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        role:
          $ref: '#/components/schemas/RoleResponse'
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    RoleResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
...
    CreateCourseRequest:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        credits:
          type: integer
    UpdateCourseRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        credits:
          type: integer
    CourseResponse:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        description:
          type: string
        credits:
          type: integer
        teachers:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        created_at:
          type: string
          format: date-time
    CourseListResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: '#/components/schemas/CourseResponse'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
    CreatePLORequest:
      type: object
      required: [code, description]
      properties:
        code:
          type: string
        description:
          type: string
    UpdatePLORequest:
      type: object
      properties:
        description:
          type: string
    PLOResponse:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
    CreateCLORequest:
      type: object
      required: [code, description]
      properties:
        code:
          type: string
        description:
          type: string
    UpdateCLORequest:
      type: object
      properties:
        description:
          type: string
    CLOResponse:
      type: object
      properties:
        id:
          type: integer
        course_id:
          type: integer
        code:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
...
    CourseAnalyticsResponse:
      type: object
      properties:
        course_id:
          type: integer
        total_students:
          type: integer
        outcome_achievement:
          type: array
          items:
            type: object
            properties:
              clo_id:
                type: integer
              achievement_rate:
                type: number
        average_grade:
          type: number
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
  responses:
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'